requirements.txt
flask==2.3.3
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.1